[project @ 2005-05-23 10:17:22 by simonpj] Fix what "standard class" means
	MERGE to STABLE

The Report's definition of defaulting (Section 4.3.4) defines a "standard class" to be
one defined by the Prelude or a standard library.  GHC was omitting the "..or a standard
library" part; and it even omitted the Prelude classes Monad and Functor.
This commit fixes the bug by adding

	Random, RandomGen,
	Monad, MonadPlus, Functor

to the classes defined as "standard".