[project @ 2005-05-26 21:37:13 by simonpj] MERGE TO STABLE
Put back in a missing case for higher-rank types. When the
definition is
	a) non-recursive
	b) a function binding
	c) lacks a type signature
we want to *infer* a perhaps-higher-rank type for the RHS,
before making a monomorphically-typed Id for the LHS.

E.g. 	f = \(x :: forall a. a->a) -> (x True, x 'c')

This case got lost in the transition to 6.4

tc194 tests it