[project @ 2005-07-12 14:50:10 by simonpj] An earlier commit, liberalising the defaulting rules for ghci, messed up the handling of defaulting at the top level.  This commit fixes it.
The current story is that if there's an ambiguous top-level
multi-parameter constraint (C a b), then a,b won't be defaulted,
regardless of how a,b are used otherwise.  A type variable is
defaulted only if it's constrained by single-parameter type classes,
even in the more-liberal GHCi.

tcfail142 tests this case.