[haddock @ 2005-08-05 07:01:12 by panne] Reverted to previous version (but with bumped version number), the last commit broke RPM building on SuSE systems due to differently named dependencies.
As a clarification: All .spec files in the repository have to work at least
on SuSE, because that's the system I'm using. And as "Mr. Building Police",
I reserve me the right to keep them that way... >:-) It might very well be
the case that we need different .spec files for different platforms, so
packagers which are unhappy with the current .spec files should contact me,
stating the actual problems.