[project @ 2005-08-11 08:04:33 by simonpj] Do 'tidying' on Kinds before printing them.  This avoids printing stuff like 'k_43b' in user error messages.
To do this, I ended up adding an OccName to Kind.KindVar.  Even
then the implementation is a bit of hack (see comments with 
Type.tidyKind).  Still, it's a highly localised hack, whereas the
"right thing" entails making KindVar into a flavour of Var, which
seems like an uncomfortably big change.

   I think this change can merge to the stable branch