[project @ 2005-11-30 14:20:06 by simonpj] ----------------------------------------- 	Fix 'mkName' operator in Template Haskell 	so that it handles built-in syntax 	-----------------------------------------
	Merge to stable branch

The 'mkName' function in Template Haskell wasn't dealing correctly with
built-in syntax.  The parser generates Exact RdrNames for built-in syntax
operators, such as ':' and '[]'; and hence so should Convert.

At the same time I'm now generating a better error message in TH when
you use a constructor as a variable or vice versa.