STM fix from Tim Harris
Fixes assertion failures with STM and -debug.  Tim says:
Sorry, it's a problem in how nested transactions are handled in
non-SMP builds.  It'll bite when trying to commit a nested transaction
which has read from a TVar but not updated it.

The call to validate_and_acquire_ownership in
stmCommitNestedTransaction should be the same as that in
stmCommitNestedTransaction, i.e.:

  result = validate_and_acquire_ownership(trec, (!use_read_phase), TRUE);