Cosmetics in SpecConstr
SpecConstr currently uses substExpr for tiresome reasons to do with
GADTs.  Unfortunately the substExpr generates some WARNINGS (when DEBUG)
is on, because we aren't adding all the in-scope Ids to the in-scope
set of the substitution.

When we move to FC these substExprs will go away, so I'm not going to
worry about this now.