Bug-fix for infix function definitions (parse/rename)
Fix a crash provoked by

	x `op` y = x
	op       = True

The trouble was that there is currently a single 'infix' flag for the
whole group; and RnTypes.checkPrecMatch was therefore expecting the
second eqn to have two args.

This fixes the crash, and also or-s the infix flags for the various
eqns together; previously it was just taken from the first eqn, which
was wrong.