Fix egregious and long-standing tidying bug
A typo in tidyAlt meant that we could get shadowing of occurrence names
in the output of tidying.  (Specifically, of existentially bound type 
variables.)  That in turn meant that an IfaceExpr could have shadowing, so
when the IfaceExpr was read in, it meant something different.
That in turn led to an obscure crash like:
	Panic: tcIfaceTyVar

Anyway, this fixes it.  MERGE into 6.4.3.