markSignalHandlers(): implementation was unnecessary, and had a bug
There's no need to mark the signal handler here, because it is stored
in a StablePtr and hence is a root anyway.  Furthermore, the call to
evac() was passing the address of a local variable, which turned out
to be harmless for copying GC, but fatal for compacting GC: compacting
GC assumes that the addresses of the roots are the same each time.

Fixes: possibly #783, possibly #776, definitely #787